Extract `DeltaQueryReceiverImpl` into its own class, refactor `InMemoryServer` to use `ParticipationManager` for improved partition and participation handling.
